Move s3 file to api dir
