+++ dluhc-component-library/src/components/siteSelectionForm/SiteSelectionForm.tsx
@@ -9,7 +9,7 @@ import { FormState, FormValue, FormPageSchema, UiSchema } from "./types";
 
 import "./SiteSelectionForm.css";
 import { createValidationSchema } from "./utils";
-import { uploadFile } from "./s3";
+import { uploadFile } from "../../api/aws/api";
 import CheckAnswers from "./components/CheckAnswers";
 
 interface SiteSelectionForm {
